# internal/server/docs/swagger.yaml
components:
  schemas:
...
    api.CollectionsCreateResponse:
      properties:
        collectionId:
          type: string
      type: object
...
    api.ErrorEnvelope:
      properties:
        error:
          $ref: '#/components/schemas/api.APIError'
      type: object
...
    api.SpaceInfo:
      properties:
        author:
          type: string
        createdAt:
          type: string
        derived:
          description: |-
            Derived marks a space created by the account's own derivation
            (POST /v1/spaces/derived/:name or another consumer of the SDK's
            Derive). Derived spaces are permanent — DELETE refuses them with
            409 space.derived_undeletable. Absent on created / joined / 1-1
            spaces.
          type: boolean
        description:
          type: string
        iconCid:
          type: string
        id:
          type: string
        name:
          type: string
        ownRole:
          enum:
          - owner
          - admin
          - writer
          - reader
          - guest
          - none
          type: string
        push:
          $ref: '#/components/schemas/api.SpacePushKeys'
        settings:
          additionalProperties: {}
          type: object
        spaceIndexObjectId:
          type: string
        spaceType:
          type: string
        status:
          type: string
        type:
          type: string
      type: object
...
    api.SpaceUpdateRequest:
      properties:
        description:
          type: string
        iconCid:
          type: string
        name:
          type: string
      type: object
...
info:
  description: Local HTTP/JSON API wrapping any-sync-sdk. Localhost-only, no auth
    in v1.
  title: Any API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /spaces/{spaceId}:
    delete:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Delete a space
      tags:
      - spaces
    get:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SpaceInfo'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Get a space
      tags:
      - spaces
    patch:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.SpaceUpdateRequest'
                description: At least one field required
                summary: body
        description: At least one field required
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Update space metadata
      tags:
      - spaces
...
  /spaces/{spaceId}/aggregate:
    post:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.SpaceAggregateRequest'
                description: Aggregation params (objectId+dataset required)
                summary: body
        description: Aggregation params (objectId+dataset required)
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.AggregateResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Aggregate over a per-object dataset
      tags:
      - data
  /spaces/{spaceId}/bundles:
    get:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.BundleListResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: List the space's bundles
      tags:
      - bundles
    post:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.BundleEnsureRequest'
                description: Bundle to register
                summary: body
        description: Bundle to register
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.BundleEnsureResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Install or adopt a bundle
      tags:
      - bundles
  /spaces/{spaceId}/bundles/{bundleId}:
    get:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Bundle ID, percent-encoded (general-chat%2Fv1)
        in: path
        name: bundleId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.BundleGetResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Read one bundle
      tags:
      - bundles
...
  /spaces/{spaceId}/bundles/{bundleId}/resolve:
    post:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Bundle ID, percent-encoded
        in: path
        name: bundleId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.BundleResolveRequest'
                description: Losing root to delete
                summary: body
        description: Losing root to delete
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Resolve a losing root
      tags:
      - bundles
  /spaces/{spaceId}/collections:
    get:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Include hidden collections
        in: query
        name: includeHidden
        schema:
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.CollectionsListResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: List collections
      tags:
      - collections
    post:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.CollectionsCreateRequest'
                description: Collection params
                summary: body
        description: Collection params
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.CollectionsCreateResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Create a collection
      tags:
      - collections
  /spaces/{spaceId}/collections/{collectionId}:
    get:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.CollectionInfo'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Get a collection
      tags:
      - collections
    patch:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.CollectionPatchRequest'
                description: Fields to change
                summary: body
        description: Fields to change
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Patch a collection's display metadata
      tags:
      - collections
  /spaces/{spaceId}/collections/{collectionId}/properties:
    get:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.PropertiesListResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: List properties of a collection
      tags:
      - collections
    post:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.AddPropertyRequest'
                description: Property draft
                summary: body
        description: Property draft
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.AddPropertyResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Add a property to a collection
      tags:
      - collections
  /spaces/{spaceId}/collections/{collectionId}/properties/{propId}:
    delete:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        schema:
          type: string
      - description: Property ID
        in: path
        name: propId
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Remove a property from a collection
      tags:
      - collections
    patch:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        schema:
          type: string
      - description: Property ID
        in: path
        name: propId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.PropertyPatchRequest'
                description: Patch
                summary: body
        description: Patch
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Patch a property of a collection
      tags:
      - collections
  /spaces/{spaceId}/datasets:
    get:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.DatasetsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: List a space's dataset schemas
      tags:
      - data
  /spaces/{spaceId}/debug:
    get:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SpaceDebugResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Debug space sync info (diagnostic, unstable)
      tags:
      - debug
  /spaces/{spaceId}/debug/objects/{objectId}:
    get:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Object ID
        in: path
        name: objectId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ObjectDebugResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Debug object tree info (diagnostic, unstable)
      tags:
      - debug
  /spaces/{spaceId}/delete-records:
    post:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.DeleteRecordsRequest'
                description: Delete params
                summary: body
        description: Delete params
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ModifyResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Delete records from a dataset
      tags:
      - data
...
  /spaces/{spaceId}/sync:
    post:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Force an immediate head-sync round (sync now)
      tags:
      - spaces
  /spaces/{spaceId}/sync-status:
    get:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SpaceSyncStatusResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Get space sync status
      tags:
      - sync
...
servers:
- url: http://127.0.0.1:7001/v1